Laravel

+++ app/Http/Controllers/FakultasController.php
@@ -29,6 +29,10 @@ class FakultasController extends Controller
      */
     public function store(Request $request)
     {
+
+       if ($request->user()->cannot('create',Fakultas::class)){
+        abort(403);
+       }
         // validasi input
         $input = $request->validate([
             'nama' => 'required|unique:fakultas',
@@ -67,6 +71,15 @@ class FakultasController extends Controller
      */
     public function update(Request $request,  $fakultas)
     {
+                $fakultas = 
+            Fakultas::findOrFail
+            ($fakultas);
+
+
+        if($request->user()->cannot('update',Fakultas::class)){
+            abort(403);
+        }
+
         $input = $request->validate([
             'nama' => 'required|unique:fakultas',
             'singkatan' => 'required|max:5',
@@ -81,12 +94,15 @@ class FakultasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($fakultas)
+    public function destroy(Request $request, $fakultas)
     {
         $fakultas = 
         Fakultas::findOrFail
         ($fakultas);
 
+        if($request->user()->cannot('delete',$fakultas)){
+            abort(403);
+        }
         $fakultas->delete();
 
     return redirect()->route('fakultas.index')->with('success','Fakultas berhasil dihapus.');
